Expense Management Web App
This project is usually meant for those who do a lot of expenses and they are unable to monitor at the end of the month they go bankrupt. So for this I made this project in which we can manage our daily expenses by enabling a budget if the budget limit exceeds then at the end of the month it will show the profit % and the loss %.There are several modules such as the set budget module, transfer amount module .The UI was designed using Figma and the frontend was implemented using HTML,CSS and javascript. The backend of the project was implemented using PHP and the database which I  used is Mysql. I used several joins in the database to interconnect different table modules. There will be no data breach because each users password is encrypted. Its quite friendly and handy though.